Extract option computation
